chore: chatApi.ts renamed to chatBotApi.ts
