+++ frontend/src/pages/ReportPage/ReportPage.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 const ReportPage = () => {
-  return <div></div>;
+  return <div>Report Page is Coming Soon.</div>;
 };
 
 export default ReportPage;
