+++ frontend/src/redux/store.ts
@@ -1,9 +1,11 @@
 import { configureStore } from '@reduxjs/toolkit';
 import exploreReducer from '@/store/slices/exploreSlice';
+import chatBotReducer from '@/store/slices/chatBotSlice';
 
 export const store = configureStore({
   reducer: {
     explore: exploreReducer,
+    chatBot: chatBotReducer,
   },
 });
